split crawler into base, web and ccnews modules
